Return since() fn from require(), instead of empty
The line `module.exports = defineSince()` was returning `undefined`
because the function wasn't returning the `since()` function, it was
only assigning to `global.since`. This meant the module's value, when
required, was an empty object.

Returning the assignment allows both assigning to `global.since` and
returning the module's value via `require()`.